requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
jupyterlab>=4.0
ipykernel>=6.0
openpyxl>=3.1
langgraph>=0.0.40
langchain>=0.2.0
google-genai>=0.5.0
python-dotenv>=1.0
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1